refactor: removed legacy
