+++ src/index.ts
@@ -5,9 +5,3 @@ export { default as Box } from './components/Box/Box';
 export { default as Button } from './components/Button/Button';
 export { default as Icon } from './components/Icon/Icon';
 export { default as Text } from './components/Text/Text';
-
-// Legacy
-export { ThemeProvider as LegacyThemeProvider } from 'styled-components/native';
-export { default as legacyTheme } from './legacy-theme';
-
-export { default as LegacyInput } from './legacy-components/Input/Input';
