--- src/legacy-components/Button/types.ts
@@ -1,5 +1,0 @@
-export type ButtonVariantTypes = 'primary' | 'secondary' | 'success' | 'danger';
-
-export type ButtonTypeTypes = 'contained' | 'outline' | 'text';
-
-export type ButtonSizeTypes = 'small' | 'medium' | 'large';
