--- src/legacy-components/Icon/list.ts
@@ -1,177 +1,0 @@
-export const iconList = [
-  'academy',
-  'account',
-  'add-note',
-  'add-photo',
-  'add-product',
-  'alert',
-  'announcement',
-  'archive',
-  'arrow-down',
-  'arrow-left',
-  'arrow-right',
-  'arrow-up',
-  'attach',
-  'automatic-pricing',
-  'award',
-  'back-fill',
-  'back',
-  'barcode',
-  'bill',
-  'book',
-  'browser',
-  'calendar-stroke',
-  'calendar',
-  'camera',
-  'campaign',
-  'change-fill',
-  'change',
-  'chart',
-  'chatbot',
-  'check-fill',
-  'check',
-  'clock',
-  'close-fill',
-  'close',
-  'coin',
-  'compass',
-  'confetti',
-  'copy',
-  'coupon',
-  'delete',
-  'dialog',
-  'dislike',
-  'document-search',
-  'document',
-  'donation',
-  'download',
-  'edit',
-  'engagement',
-  'excel',
-  'exit',
-  'expand',
-  'export',
-  'external-link',
-  'external-share',
-  'eye-off',
-  'eye-on',
-  'facebook',
-  'feedback',
-  'filter',
-  'finance-accounting',
-  'fire',
-  'flash-fill',
-  'flash',
-  'forward',
-  'gift',
-  'globe',
-  'go-forward',
-  'graphic',
-  'growth',
-  'hamburger-menu',
-  'hamburger',
-  'handshake',
-  'happy',
-  'heart-line',
-  'heart',
-  'help',
-  'home',
-  'image',
-  'influencer',
-  'info',
-  'label',
-  'light-bulb',
-  'like',
-  'link',
-  'listing',
-  'live-monitor',
-  'live-support',
-  'loading',
-  'lock',
-  'magic-wand',
-  'mail-opened',
-  'mail',
-  'map',
-  'market-order-back',
-  'market-order-cancel',
-  'market-order-change',
-  'market-order-check',
-  'market',
-  'message',
-  'metric-status-down',
-  'metric-status-up',
-  'minus-fill',
-  'minus',
-  'mobile-filled',
-  'mobile-settings',
-  'mobile-stroke',
-  'money-transfer',
-  'motorcycle',
-  'my-offers',
-  'normal',
-  'notification',
-  'order-back',
-  'order-box',
-  'order-cancel',
-  'order-check',
-  'order-return',
-  'order-tracking',
-  'order',
-  'paper',
-  'pause-fill',
-  'pause',
-  'people',
-  'phone-settings',
-  'phone',
-  'photo',
-  'pin',
-  'play-fill',
-  'play',
-  'plus-fill',
-  'plus',
-  'preview',
-  'print',
-  'product',
-  'profile-check',
-  'profile',
-  'rating',
-  'report',
-  'rotate',
-  'sad',
-  'save',
-  'search',
-  'seller-ads',
-  'send-mail',
-  'send',
-  'services',
-  'settings',
-  'share',
-  'shopping-bag-back',
-  'shopping-bag-cancel',
-  'shopping-bag-discount',
-  'shopping-bag-return',
-  'shopping-bag',
-  'sorting',
-  'sound-off',
-  'sound-on',
-  'star',
-  'stop-fill',
-  'stop',
-  'store-performance',
-  'store',
-  'suitcase',
-  'support',
-  'ticket',
-  'truck',
-  'turn-back',
-  'unlock',
-  'upload',
-  'user-management',
-  'vacation-mode',
-  'wallet',
-  'warning',
-  'web-filled',
-  'web-stroke',
-  'zoom-in',
-  'zoom-out',
-] as const;
