--- src/legacy-components/Text/types.ts
@@ -1,28 +1,0 @@
-export type TextVariantTypes =
-  | 'heading1'
-  | 'heading2'
-  | 'heading3'
-  | 'subtitle01Regular'
-  | 'subtitle01Medium'
-  | 'subtitle01Semibold'
-  | 'subtitle01Bold'
-  | 'subtitle02Regular'
-  | 'subtitle02Medium'
-  | 'subtitle02Semibold'
-  | 'subtitle02Bold'
-  | 'subtitle03Regular'
-  | 'subtitle03Medium'
-  | 'subtitle03Semibold'
-  | 'subtitle03Bold'
-  | 'subtitle04Regular'
-  | 'subtitle04Medium'
-  | 'subtitle04Semibold'
-  | 'subtitle04Bold'
-  | 'bodyText'
-  | 'bodyUnderline'
-  | 'bodyTextLink'
-  | 'bodyLongText'
-  | 'captionText'
-  | 'captionMedium'
-  | 'captionLongText'
-  | 'captionTextLink';
